Repackaging and improving tests
