--- magnet/src/main/java/magnet/inspection/Instance.java
@@ -1,25 +1,0 @@
-package magnet.inspection;
-
-import magnet.Scoping;
-import org.jetbrains.annotations.NotNull;
-
-public interface Instance {
-
-    enum Provision { BOUND, INJECTED }
-
-    @NotNull
-    Scoping getScoping();
-
-    @NotNull
-    String getClassifier();
-
-    @NotNull
-    Class<?> getType();
-
-    @NotNull
-    Object getValue();
-
-    @NotNull
-    Provision getProvision();
-
-}
