+++ magnet/src/test/java/magnet/internal/events/OnExitScope.java
@@ -1,4 +1,4 @@
-package magnet.inspection.events;
+package magnet.internal.events;
 
 import magnet.Scope;
