+++ magnetx-app-stetho-scope/src/main/java/magnetx/app/stetho/scope/ScopeDumper.kt
@@ -1,12 +1,27 @@
+/*
+ * Copyright (C) 2019 Sergej Shafarenka, www.halfbit.de
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package magnetx.app.stetho.scope
 
 import magnet.Classifier
 import magnet.Scope
-import magnet.inspection.Instance
-import magnet.inspection.ScopeVisitor
+import magnet.Scope.Visitor.Instance
 import java.io.PrintStream
 
-internal class ScopeDumper(private val writer: PrintStream) : ScopeVisitor {
+internal class ScopeDumper(private val writer: PrintStream) : Scope.Visitor {
 
     private val instances = mutableListOf<Instance>()
     private var currentScope: Scope? = null
